Added in line numbers
This is my own personal fix for line numbers. In editorUpdateRows it already keeps track of the row number, so I print it. In editorRefreshScreens, it uses snprintf as a way to get the length of the cursor position plus one as a string (it's zero indexed) to correct the offset from printing the numbers.